requirements.txt
absl-py==0.11.0
albumentations==1.0.0
altair==4.1.0
antlr4-python3-runtime==4.8
anykeystore==0.2
appdirs==1.4.4
argon2-cffi==20.1.0
astor==0.8.1
astunparse==1.6.3
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
base58==2.1.0
behave==1.2.6
black==21.4b2
bleach==3.3.0
blinker==1.4
bravado==11.0.3
bravado-core==5.17.0
cachetools==4.2.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
clang==5.0
click==7.1.2
cloudpickle==1.6.0
coloredlogs==15.0
configparser==5.0.2
coolname==1.1.0
cryptacular==1.5.5
cycler==0.10.0
Cython==3.0a6
decorator==4.4.2
defusedxml==0.6.0
docker-pycreds==0.4.0
easydict==1.9
efficientnet-pytorch==0.6.3
entrypoints==0.3
filelock==3.0.12
flake8==3.8.4
flatbuffers==1.12
future==0.18.2
gast==0.4.0
gdown==3.12.2
gitdb==4.0.7
GitPython==3.1.14
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
GPUQueue==0.0.3
grpcio==1.40.0
h5py==3.1.0
humanfriendly==9.1
hupper==1.10.2
hydra-core==1.1.0
idna==2.10
imageio==2.9.0
imgaug==0.4.0
iniconfig==1.1.1
iopath==0.1.9
ipdb==0.13.4
ipykernel==5.4.3
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.17.2
Jinja2==3.0.1
joblib==1.0.0
jsonpointer==2.1
jsonref==0.2
jsonschema==3.2.0
jupyter-client==6.1.11
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
linecache2==1.0.0
Markdown==3.3.3
MarkupSafe==2.0.1
matplotlib==3.3.3
mccabe==0.6.1
MedPy==0.4.0
mistune==0.8.4
monotonic==1.6
msgpack==1.0.2
munch==2.5.0
mypy-extensions==0.4.3
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.5
nibabel==3.2.1
nilearn==0.8.0
notebook==6.4.0
numpy==1.19.4
oauthlib==3.1.0
omegaconf==2.1.0
opencv-python==4.5.1.48
opencv-python-headless==4.5.1.48
opt-einsum==3.3.0
packaging==20.9
pandas==1.1.5
pandocfilters==1.4.3
parse==1.19.0
parse-type==0.5.2
parso==0.7.1
PasteDeploy==2.1.1
pathspec==0.9.0
pathtools==0.1.2
pbkdf2==1.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
plaster==1.0
plaster-pastedeploy==0.7
plotly==5.1.0
pluggy==1.0.0.dev0
portalocker==2.2.1
pretrainedmodels==0.7.4
prometheus-client==0.10.1
promise==2.3
prompt-toolkit==3.0.8
protobuf==3.14.0
psutil==5.8.0
ptyprocess==0.6.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.2
pycodestyle==2.6.0
pycparser==2.20
pydeck==0.6.2
pydot==1.4.1
pyflakes==2.2.0
Pygments==2.7.3
PyJWT==2.0.1
pynrrd==0.4.2
pyparsing==2.4.7
pyquaternion==0.9.9
pyramid==1.10.5
pyramid-mailer==0.15.1
pyrsistent==0.17.3
PySocks==1.7.1
pytest==6.2.4
pytest-remotedata==0.3.1
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.5
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==22.0.2
regex==2021.8.3
repoze.sendmail==4.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rfc3987==1.3.8
rsa==4.6
runx==0.0.10
scikit-image==0.18.0
scikit-learn==0.24.2
scipy==1.5.4
Send2Trash==1.5.0
sentry-sdk==1.1.0
Shapely==1.7.1
shortuuid==1.0.1
SimpleITK==2.0.2
simplejson==3.17.2
six==1.15.0
sklearn==0.0
smmap==4.0.0
SQLAlchemy==1.3.23
streamlit==0.82.0
strict-rfc3339==0.7
subprocess32==3.5.4
swagger-spec-validator==2.7.3
tabulate==0.8.7
tenacity==8.0.1
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.7.0
tensorboardX==2.1
tensorflow==2.6.0
tensorflow-estimator==2.6.0
termcolor==1.1.0
terminado==0.10.0
terminaltables==3.1.0
testpath==0.5.0
threadpoolctl==2.1.0
tifffile==2020.12.8
timm==0.3.2
toml==0.10.2
toolz==0.11.1
torch==1.7.1+cu110
torchvision==0.8.2+cu110
tornado==6.1
tqdm==4.54.1
traceback2==1.4.0
traitlets==5.0.5
transaction==3.0.1
translationstring==1.4
typing==3.7.4.3
typing-extensions==3.7.4.3
tzlocal==2.1
unittest2==1.1.0
urllib3==1.26.6
validators==0.18.2
velruse==1.1.1
venusian==3.0.0
wandb==0.11.0
watchdog==2.1.2
wcwidth==0.2.5
webcolors==1.11.1
webencodings==0.5.1
WebOb==1.8.6
websocket-client==0.58.0
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
WTForms==2.3.3
wtforms-recaptcha==0.3.2
yacs==0.1.8
zope.deprecation==4.4.0
zope.interface==5.2.0
zope.sqlalchemy==1.3
